Basic JavaScript
You can use the scripting language JavaScript to add interactive elements to web pages. Along with HTML and CSS, it is one of the fundamental web technologies and is compatible with all current browsers.

You will study the fundamentals of JavaScript programming in this course. Basic data structures like strings and numbers will be your first to work with. After that, you'll learn how to work with loops, if/else expressions, functions, objects, arrays, and more.